+++ lib/dbas_filesystem.dart
@@ -1,1 +1,8 @@
 export 'src/dbas_filesystem.dart' show DbasFileSystem;
+export 'src/dbas_filesystem_exceptions.dart'
+    show
+        DbasFileSystemException,
+        FileNotFoundException,
+        DirectoryNotFoundException,
+        FileAlreadyExistsException,
+        DirectoryNotEmptyException;